test: extract graph-only config regression to dedicated class
Keep ModuleGraphPluginFunctionalTest under detekt LargeClass after
the issue #70 functional coverage was added.
